backend/requirements.txt
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
billiard==4.2.1
blinker==1.9.0
boto3==1.38.18
botocore==1.38.18
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
docx==0.2.4
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
kombu==5.5.3
lxml==5.4.0
Mako==1.3.10
MarkupSafe==3.0.2
openai==1.79.0
packaging==25.0
pika==1.3.2
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.6
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
redis==6.1.0
requests==2.32.3
s3transfer==0.12.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
stripe==12.1.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
